requirements.txt
sdnotify==0.3.2
paho-mqtt==1.4.0